% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mdi.adj.R
\name{calculate.mdi.adj}
\alias{calculate.mdi.adj}
\title{Calculate Microbial Dysbiosis Index (MDI, Gevers 2014)}
\usage{
calculate.mdi.adj(dat, depth)
}
\arguments{
\item{depth}{rarefaction depth (uniform sample depth)}

\item{datobj}{a matrix of taxa relative abundances; sequencing sample identifiers are in row.names.}
}
\value{
a numerical vector of MDI values
}
\description{
Calculate Microbial Dysbiosis Index (log ratio of sum RAs of
taxa incraesed in IBD/sum RAs of taxa decraesed in IBD); from Gevers 2014
Bit note R. gnavus moved to "bad side" (nominator)
}
