% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.to.date.R
\name{Change.to.date}
\alias{Change.to.date}
\title{Correctly Format Metadata Dates}
\usage{
Change.to.date(keyobj, in.format = "\%d/\%m/\%Y")
}
\arguments{
\item{keyobj}{input metadata dataframe}

\item{informat}{format of input chracter string, defalt is "%d/%m/%Y"}
}
\value{
key with all date columns formatted as Date
}
\description{
Change metadata columns which have "date" in name  to 'date' format
}
