% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.chooser.R
\name{color.chooser}
\alias{color.chooser}
\title{Show Available Hues of Color}
\usage{
color.chooser(x)
}
\arguments{
\item{name}{of a color}
}
\value{
a piechart of avaialble colors
}
\description{
Show available hues for any particular color-helps choose colors for plots!
}
