% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_code_taxa.R
\name{color_code_taxa}
\alias{color_code_taxa}
\title{Assign Specific Colors to Lineages}
\usage{
color_code_taxa(tax, dattable)
}
\arguments{
\item{tax}{a vector of full-lineage taxa names}

\item{dattable}{a datobj object}
}
\value{
a named character vector of color names; taxa names are the names of the vector
}
\description{
Assign color to taxa based on phylogenetic lineage and according to a specific 
pre-loaded taxa to colors key- this must be loaded into your environment and called
`cod`.
}
