% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.1dim.r
\name{cor.1dim}
\alias{cor.1dim}
\title{Make  a one-dimensional correlation table}
\usage{
cor.1dim(datobj, yfv, dr, use = "complete", method = "spearman", ...)
}
\arguments{
\item{datobj}{a matrix/dataframe of numerical data, usually taxa RAs}

\item{yfv}{vector of interest}

\item{dr}{num digits to round after 0}
}
\value{
table of all correlations, giving R, p abd q
}
\description{
Correlate a single vector against several vectors; spearmans by default
Order of samples MUST BE SAME
}
