% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.2dim.R
\name{cor.2dim}
\alias{cor.2dim}
\title{Make a 2-dimensional correlation table}
\usage{
cor.2dim(x.cors, y.cors, dig, cut.level = "no")
}
\arguments{
\item{x.cors}{a matrix/data frame  of numerical vectors}

\item{y.cors}{a matrix/data frame  of numerical vectors; sample order same as x.cors}

\item{dig}{num digits to round after 0}

\item{cut.level}{if set to "yes", will strip taxa names to genus level}
}
\value{
table of all correlations, giving R, p and q
}
\description{
Correlate several vector against several vectors, spearmans by default
ORDER OF SAMPLES MUST BE SAME IN `x` and `y``
}
