% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dat_functions.R
\name{filt_data}
\alias{filt_data}
\title{Filter  Data}
\usage{
filt_data(dat.lis, x, tmax)
}
\arguments{
\item{dat.lis}{perfectly matched list of key and dat (column as ASVs, raw abundance)}

\item{x}{prevalence threshold;features whsoe prevalence is equal or smaller than x 
will be dropped}

\item{tmax}{minimal RA/abundance threshold; features whose maximal RA/abundance is lower 
than this are dropped (i.e., must pass this threshold in at least ONE sample to be 
retained). Note this step is run AFTER the prevalence filtration.}
}
\value{
dat.lis where the data object is filtered from rare ASVs
}
\description{
Remove rare ASvs based on prevalence and minimal abundance
}
