% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtCols.R
\name{filtCols}
\alias{filtCols}
\title{Filter featres by prevalence}
\usage{
filtCols(dat, x)
}
\arguments{
\item{x}{prevalence threshold;features whose pervalence is equal or smaller than x 
will be dropped}

\item{datobj}{matrix of numerical data}
}
\value{
filtered matrix
}
\description{
FIlter taxa based on prevalence only 
(sampels with prevalence higher than `x` are retained )
}
