% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtColsTmax.R
\name{filtColsTmax}
\alias{filtColsTmax}
\title{Feature filtration by prevalence and min abundance}
\usage{
filtColsTmax(datobj, x, tmax)
}
\arguments{
\item{datobj}{matrix of numerical data}

\item{x}{prevalence threshold;features whocer pervalence is equal or smaller tahn x 
will be dropped}

\item{tmax}{minimal RA/abundance threshold; features whose maximal RA/abundance is lower 
than this are dropped (i.e., must pass this threshold in at least ONE sample to be 
retained). Note this step is run AFTER the prevalence filtration.}
}
\value{
filtered matrix
}
\description{
FIlter taxa based on prevalence and on an abundance threshold
}
