% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_cooc.R
\name{get_cooc_stats}
\alias{get_cooc_stats}
\title{Get coccourence Stats}
\usage{
get_cooc_stats(asvx, dfall)
}
\arguments{
\item{asvx}{ASV i want to get stats on}

\item{dfx}{a dataframe with all proportion and counts of each ASV across sample types. 
each row represents one depth in one iteratio for one ASV}
}
\value{
dfres aa datafreme , per ASV , giving: lowest depth it was found in, number depths it was found in, 
num times it was found (all iterations on all dephts), mean RA and mean prop for fecal and oral. Note those are calculated only
from rows wehre it was found (i.e., number fo depths/itetations it was not found dont play a part)
}
\description{
Get Stats per ASV across rarefaction depths and iterations
}
