% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dat_functions.R
\name{get_data}
\alias{get_data}
\title{Read In SeqTable and Metadata Key}
\usage{
get_data(pathToKey, pathToDat)
}
\arguments{
\item{pathToKey}{pathway to metadata key (sequencing sample identifers in
column "ID", samples in rows)}

\item{pathToDat}{pathway to abundance data ( sequencing sample identifers
in row.names)}
}
\value{
list of key and dat, properly matched; prints out to console sample numbers;
warns if sample names dont match. If dat has sampels missing from key, they will be dropped.
}
\description{
Read in a dada seqtable (with columns labelled as ASVs rather than raw sequences)
and corresponding key; match data and key (same samples in same order for both objects)
}
