% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_loop.R
\name{lm_loop}
\alias{lm_loop}
\title{Build linear model per feature}
\usage{
lm_loop(
  keyobj,
  datobj,
  fixed.vars,
  vars.to.scale = "none",
  col.merge = "ID",
  transform = "log2"
)
}
\arguments{
\item{keyobj}{a dataframe of metadata; sequencing sample identifiers in collumn "ID"}

\item{datobj}{a matrix of taxa absolute/relative abundances; sequencing sample identifiers are in row.names. Rare taxa should be removed.}

\item{fixed.vars}{character vector of predictors for model (names of relevant columns in keyobj)}

\item{vars.to.scale}{chracter vector of names of fixed.vars that need to be scales beforemodelling (usually any numerical variable)}

\item{col.merge}{column to merge on , default is "ID}
}
\value{
table of all coefficnts for each microbial feature
}
\description{
Run linear model for each microbial taxon (like masslin but better)
}
