% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_loop.R
\name{lmer_loop}
\alias{lmer_loop}
\title{Run mixed linear model with random effect
for each microbial taxon (like masslin but better)}
\usage{
lmer_loop(
  keyobj,
  datobj,
  transform = "log2",
  fixed.vars,
  vars.to.scale = "none",
  random.var
)
}
\arguments{
\item{keyobj}{a dataframe of metadata; sequencing sample identifiers in collumn "ID"}

\item{datobj}{a matrix of taxa absolute/relative abundances; sequencing sample identifiers are in row.names. Rare taxa should be removed.}

\item{fixed.vars}{character vector of predictors for model (names of relevant columns in keyobj)}

\item{vars.to.scale}{chracter vector of names of fixed.vars that need to be scales beforemodelling (usually any numerical variable)}

\item{random.var}{variable to be used for random effect}
}
\value{
table of all coefficnts for each microbial feature
}
\description{
Run mixed linear model with random effect
for each microbial taxon (like masslin but better)
}
