% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.pcoa.R
\name{make.pcoa}
\alias{make.pcoa}
\title{Draw a PCoA}
\usage{
make.pcoa(datobj, keyobj, kind, dis, var1, var2, name1, name2, i, j, ...)
}
\arguments{
\item{datobj}{matrix of RA/abundance data (numeric), or a Distance matrix}

\item{keyobj}{correspondin metadata key; sampels must be in same order!}

\item{kind}{type of data - `unifrac` (distance matrixes) or `taxtable` (ASVs count as taxtables )}

\item{dis}{distance index -  `bray` or  `jaccard` - relevant only if `kind` is taxtable}

\item{var1}{variable of interest  -give actual values (i.e., key$group)}

\item{var2}{optional - anotehr variable of interest  -give actual values (i.e., key$Group)}

\item{name1}{name of var1 (i.e., "group")}

\item{name2}{name of var2 (optional)}

\item{i}{number of coordinate to plot as X}

\item{j}{number of coortdinate to plot as y}
}
\value{
PCoA plot withn ANOSIM statistics shown
}
\description{
Make a PCoA,map dot color (and optional, shape ) to a metadata factor. If data
inputted is a a RA/abundance table, it will be converted to distance object using vegdist
}
