% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cross_sectional.R
\name{make_cross_section}
\alias{make_cross_section}
\title{Make a cross-section of longitudinal data}
\usage{
make_cross_section(keyobj, index)
}
\arguments{
\item{keyobj}{metadata key; subject column called "PatID"}

\item{index}{"first" or "last" - specify which  sample should be taken. This 
only works if teh key is ordered by date first!}
}
\value{
a cross sectional metadata key
}
\description{
Make a longitudinal metadata key into a cross sectional one - by choosing
one sample per subject.NOTE: Function assumes the column with subject identity is
called "PatID", and that it is sorted by date
}
