% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unifracs.R
\name{make_unifrac}
\alias{make_unifrac}
\title{Make UniFrac matrices}
\usage{
make_unifrac(
  taxobj,
  rawdatobj = rarified.lis[[2]],
  outgroup = "Methano",
  pathToOutdir = NULL
)
}
\arguments{
\item{outgroup}{partial name of phylogentic group to root on; "Methano" selected by default. If more than 1 ASV is 
found in this taxonomial group, the first one will be used as the outgroup}

\item{pathToOutdir}{optional - specifcy path to directory in which unifrac weighted and unwieghted matrices will be written}

\item{taxa}{dataframe of asvID and tax affilations}

\item{datobj}{dataframe of ASVs abundances, alraedy filtered and rariefied}
}
\value{
a list of weighted and unweighetd UniFrac matrices
}
\description{
Make abundance weighted/unweighted  distance matrices using UniFrac method
}
