% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine_gnavus.R
\name{mine.gnavus.1}
\alias{mine.gnavus.1}
\title{Identify  R.gnavus Sequences}
\usage{
mine.gnavus.1(taxaobj, path.out)
}
\arguments{
\item{taxaobj}{taxa table}

\item{path}{to write output}
}
\description{
This taxonomy is not given automatically by dada2. this function finds all sequnences which may be 
R.gnavus (Unassigned Lachnospiraceae) and writes them out as a fasta file which can be explored by 
blast to which ASVs of Unassigned Lachnospiraceaeare are actually  R. gnavus
}
