% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dat_key.R
\name{mrdk}
\alias{mrdk}
\title{Make Dataframe of Metadata and Microbiome Data}
\usage{
mrdk(keyobj, datobj)
}
\arguments{
\item{keyobj}{a dataframe of metadata; sequencing sample identifiers in collumn "ID"}

\item{datobj}{a matrix of taxa absolute/relative abundances; sequencing sample identifiers are in row.names}
}
\value{
a dataframe with seqeunce identifiers in "ID" column
}
\description{
Combine taxa abundances/relative abundances and metadata
}
