% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_lefse.R
\name{prep.lefse}
\alias{prep.lefse}
\title{Make LEfSe Input Table}
\usage{
prep.lefse(keyobj, datobj, varname, path, dirname)
}
\arguments{
\item{keyobj}{a dataframe of metadata; sequencing sample identifiers in collumn "ID"}

\item{datobj}{a matrix of taxa absolute/relative abundances; sequencing sample identifiers are in row.names. Rare taxa should be removed.}

\item{varname}{varaible to compare groups on}

\item{dirname}{directory of comparison}
}
\description{
Make table formatted to serve as input for a LEfSe analysis, comparing 2 groups based on
one metadata factor
}
