% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dat_functions.R
\name{rarefy_data}
\alias{rarefy_data}
\title{Rarify data}
\usage{
rarefy_data(dat.lis, depth)
}
\arguments{
\item{dat.lis}{perfectly matched list of key and dat (column as ASVs, raw abundance)}

\item{depth}{data wiull be rarieifed to this depth; samples with fewer seqeunes will be
dropped from boith dat and key; thsi will be reported in console}
}
\value{
dat.lis where the data object is filtered from rare ASVs
}
\description{
Rarefy data to a common depth. remove samples not reaching this depth from both dat and key
}
