% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_names_silva_dada.R
\name{strip.names.silva.dada}
\alias{strip.names.silva.dada}
\title{Shorten Full-Taxonomy Names}
\usage{
strip.names.silva.dada(x, t)
}
\arguments{
\item{x}{vector of taxa names that needs to be cut}

\item{t}{taxonmic rank of interest: 1 is phyla, 2 order, and so on}
}
\value{
vector of names cut down to rank of interest
}
\description{
Cut taxa names to required phylogebnetic level only.
NOTE:Taxa that are not assigned down to the required level will get be assigned the
 deepest level of taxonomny they DO havE! Special character in taxa names are changed 
 to dors or underscores
}
