% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_names_silva_fungal_fixed.R
\name{strip.names.silva.fungal.fixed}
\alias{strip.names.silva.fungal.fixed}
\title{Shorten Full-Taxonomy Eukaryal Names}
\usage{
strip.names.silva.fungal.fixed(x, t)
}
\arguments{
\item{x}{vector of taxa names that needs to be cut}

\item{t}{taxonmic rank of interest: 1 is phyla, 2 order, and so on}
}
\value{
vector of names cut down to rank of interest
}
\description{
Cut taxa names to required phylogenetic level only.
NOTE:Taxa that are not assigned down to the required level will be assigned the
 deepest level of taxonomny they DO have! Special character in taxa names are
  changed to dots or underscores
}
