% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_dat.R
\name{subdat}
\alias{subdat}
\title{Match Microbiome and Metadata}
\usage{
subdat(keyobj, datobj, kind = "tax.table")
}
\arguments{
\item{keyobj}{metadata key; sampels in rows;  column of sample sequncing identifiers
called "ID"}

\item{datobj}{numerical matrix or dataframe; Default is 'tax.table'  - a RA
or abundance table with samples in rows; can also accept UniFrac distance matrix.
Sample seqeunce identifiers must be in rownames (and if its UniFrac,also in colnames)}

\item{kind}{kind of data input - RA/abundance table ('tax.table' , default) or UniFrac
distance matrix ('unifrac')}
}
\value{
a data object perfectly matched to key
}
\description{
Matches data (datobj) and metadata (keyobj) so both have same samples in same order
If input data is a distance matrix, will match both rows and columns
}
