% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dat_functions.R
\name{sum_duplicates}
\alias{sum_duplicates}
\title{Sum Duplicates}
\usage{
sum_duplicates(dat.lis, sample.id = "SampleID")
}
\arguments{
\item{dat.lis}{A list of key and dat,perfectly matched; may be prepared with
get_data()}
}
\value{
A list of key, and taht, still perfectly matched, but abundances of multiple
runs of same sampel will be summed to one entry
}
\description{
Sum duplicates, if exists (samples which were sequenced in more than one run. They will have
same sample name , but unique sample sequencing
identifiers (ID column in key))
}
