% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dat_functions.R
\name{tax.seq.table}
\alias{tax.seq.table}
\title{Summarize by Taxonomy}
\usage{
tax.seq.table(dat.lis = rariefied.lis, taxobj, tax.level = "genus")
}
\arguments{
\item{dat.lis}{perfectly matched list of key and dat; by default after filtartion and rarefaction}

\item{taxobj}{taxa table  prepared with get_taxa()}

\item{tax.level}{taxonomic level required; default is genus}
}
\value{
dat.lis with teh dat object summarized to phylogebtic rtank of interest
}
\description{
Summarize ASV table to required phylogenetic level (default is genus)
}
