% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_dat.R
\name{td}
\alias{td}
\title{Transform Data}
\usage{
td(trans.dat, impute.method = "half-min", transform.method = "log")
}
\arguments{
\item{trans.dat}{numerical data which needs to be transformed; may be matrix, 
dataframe or vector}

\item{impute.method}{methos for zero imputation - by default, half teh minimal
value for that feature}

\item{transform.method}{method for transformation (log, log2,log10 or arcsine)}
}
\value{
transformed data
}
\description{
Apply a transformation to data; zero-impute first if necessary
}
