% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tax_tables.R
\name{write.tax.tables}
\alias{write.tax.tables}
\title{Write Summarized Taxonomy RA Tables}
\usage{
write.tax.tables(
  taxobj,
  datobj = rarified.lis[[2]],
  strip = "no",
  transpose = "no",
  phylo.levels = c("phylum", "order", "family", "genus", "species"),
  pathToOutdir
)
}
\arguments{
\item{strip}{strip taxa names to one loevel (default: no, set "yes" to strip)}

\item{transpose}{transpose data (default: no, set "yes" to transpose)}

\item{phylo.levels}{optional; specify phyl levels to write out; by defualt all will be selected}

\item{pathToOutdir}{specify path to directory in which taxonomy tables will be written}

\item{taxa}{dataframe of asvID and tax affilations}

\item{rawdat.ra}{dataframe of ASVs abundances, alraedy filtered and rariefied}
}
\description{
write out taxonomy data tables summarized for each  taxonomy rank (phyla to species)
}
